feat: 完善用户角色与权限管理

+++ kubemanage-another/controller/user/user.go
@@ -1,9 +1,11 @@
 package user
 
 import (
+	"fmt"
+	"strconv"
+
 	"github.com/gin-gonic/gin"
 	"github.com/noovertime7/kubemanage/pkg/core/kubemanage/v1"
-	"strconv"
 
 	"github.com/noovertime7/kubemanage/dto"
 	"github.com/noovertime7/kubemanage/middleware"
@@ -11,6 +13,77 @@ import (
 	"github.com/noovertime7/kubemanage/pkg/globalError"
 	"github.com/noovertime7/kubemanage/pkg/utils"
 )
+
+func (u *userController) List(c *gin.Context) {
+	page, _ := strconv.Atoi(c.DefaultQuery("page", "1"))
+	size, _ := strconv.Atoi(c.DefaultQuery("pageSize", "20"))
+	out, e := v1.CoreV1.System().User().List(c, page, size, c.Query("keyword"))
+	if e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ServerError, e))
+		return
+	}
+	middleware.ResponseSuccess(c, out)
+}
+func (u *userController) Create(c *gin.Context) {
+	var in dto.UserCreateInput
+	if e := c.ShouldBindJSON(&in); e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ParamBindError, e))
+		return
+	}
+	out, e := v1.CoreV1.System().User().Create(c, &in)
+	if e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ServerError, e))
+		return
+	}
+	middleware.ResponseSuccess(c, out)
+}
+func (u *userController) Update(c *gin.Context) {
+	id, e := strconv.Atoi(c.Param("id"))
+	if e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ParamBindError, e))
+		return
+	}
+	var in dto.UserUpdateInput
+	if e = c.ShouldBindJSON(&in); e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ParamBindError, e))
+		return
+	}
+	claims := utils.GetUserInfo(c)
+	if claims != nil && claims.ID == id {
+		if (in.Enable != nil && *in.Enable != 1) || (in.AuthorityId != 0 && in.AuthorityId != claims.AuthorityId) {
+			middleware.ResponseError(c, globalError.NewGlobalError(globalError.AuthErr, fmt.Errorf("不能冻结当前用户或修改当前用户角色")))
+			return
+		}
+	}
+	if e = v1.CoreV1.System().User().Update(c, id, &in); e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ServerError, e))
+		return
+	}
+	middleware.ResponseSuccess(c, "操作成功")
+}
+func (u *userController) Enable(c *gin.Context) {
+	id, e := strconv.Atoi(c.Param("id"))
+	if e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ParamBindError, e))
+		return
+	}
+	var in struct {
+		Enable int `json:"enable"`
+	}
+	if e = c.ShouldBindJSON(&in); e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ParamBindError, e))
+		return
+	}
+	if claims := utils.GetUserInfo(c); claims != nil && claims.ID == id && in.Enable != 1 {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.AuthErr, fmt.Errorf("不能冻结当前登录用户")))
+		return
+	}
+	if e = v1.CoreV1.System().User().SetEnable(c, id, in.Enable); e != nil {
+		middleware.ResponseError(c, globalError.NewGlobalError(globalError.ServerError, e))
+		return
+	}
+	middleware.ResponseSuccess(c, "操作成功")
+}
 
 // Login godoc
 // @Summary 管理员登录
@@ -61,10 +134,15 @@ func (u *userController) LoginOut(ctx *gin.Context) {
 	// 从context中，取出claims的值。如果不存在，则说明没有这个用户的登录信息，报错内部错误
 	claims, exists := ctx.Get("claims")
 	if !exists {
-		v1.Log.Error(globalError.ServerError)
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.AuthorizationError, fmt.Errorf("登录信息不存在")))
+		return
 	}
 	// 断言claims是否为CustomClaims类型，转成CustomClaims类型
-	cla, _ := claims.(*pkg.CustomClaims)
+	cla, ok := claims.(*pkg.CustomClaims)
+	if !ok || cla == nil {
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.AuthorizationError, fmt.Errorf("登录信息无效")))
+		return
+	}
 	// 使用全局的v1.CoreV1对象，获取一个SystemInterface接口的对象，然后获取一个UserService对象，调用LoginOut方法
 	if err := v1.CoreV1.System().User().LoginOut(ctx, cla.ID); err != nil {
 		// 登出失败，日志记录错误信息
@@ -100,6 +178,7 @@ func (u *userController) GetUserInfo(ctx *gin.Context) {
 	if err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
 	}
 	middleware.ResponseSuccess(ctx, userInfo)
 }
@@ -118,6 +197,7 @@ func (u *userController) SetUserAuthority(ctx *gin.Context) {
 	if err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
 	}
 	params := &dto.SetUserAuth{}
 	if err := params.BindingValidParams(ctx); err != nil {
@@ -125,22 +205,15 @@ func (u *userController) SetUserAuthority(ctx *gin.Context) {
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
 		return
 	}
+	if claims := utils.GetUserInfo(ctx); claims != nil && claims.ID == uid && claims.AuthorityId != params.AuthorityId {
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.AuthErr, fmt.Errorf("不能修改当前登录用户的角色")))
+		return
+	}
 	if err := v1.CoreV1.System().User().SetUserAuth(ctx, uid, params.AuthorityId); err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
 		return
 	}
-	// token中存在角色信息，需要生成新的token
-	claims := utils.GetUserInfo(ctx)
-	claims.AuthorityId = params.AuthorityId
-	newToken, err := pkg.JWTToken.GenerateToken(claims.BaseClaims)
-	if err != nil {
-		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
-		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
-		return
-	}
-	ctx.Header("new-token", newToken)
-	ctx.Header("new-expires-at", strconv.FormatInt(claims.ExpiresAt, 10))
 	middleware.ResponseSuccess(ctx, "操作成功")
 }
 
@@ -157,6 +230,11 @@ func (u *userController) DeleteUser(ctx *gin.Context) {
 	if err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
+	}
+	if claims := utils.GetUserInfo(ctx); claims != nil && claims.ID == uid {
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.AuthErr, fmt.Errorf("不能删除当前登录用户")))
+		return
 	}
 	if err := v1.CoreV1.System().User().DeleteUser(ctx, uid); err != nil {
 		v1.Log.ErrorWithCode(globalError.DeleteError, err)
@@ -179,6 +257,12 @@ func (u *userController) ChangePassword(ctx *gin.Context) {
 	if err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
+	}
+	claims := utils.GetUserInfo(ctx)
+	if claims == nil || (claims.ID != uid && claims.AuthorityId != pkg.AdminDefaultAuth) {
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.AuthErr, fmt.Errorf("只能修改自己的密码")))
+		return
 	}
 	params := &dto.ChangeUserPwdInput{}
 	if err := params.BindingValidParams(ctx); err != nil {
@@ -206,8 +290,14 @@ func (u *userController) ResetPassword(ctx *gin.Context) {
 	if err != nil {
 		v1.Log.ErrorWithCode(globalError.ParamBindError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
 	}
-	if err := v1.CoreV1.System().User().ResetPassword(ctx, uid); err != nil {
+	params := &dto.ResetUserPwdInput{}
+	if err := params.BindingValidParams(ctx); err != nil {
+		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ParamBindError, err))
+		return
+	}
+	if err := v1.CoreV1.System().User().ResetPassword(ctx, uid, params.NewPwd); err != nil {
 		v1.Log.ErrorWithCode(globalError.ServerError, err)
 		middleware.ResponseError(ctx, globalError.NewGlobalError(globalError.ServerError, err))
 		return
